update_env_file can now edit any environment variable (Also supported using an arbitrary logger in these functions)
